# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/dashboard:
    get:
      operationId: getDashboardMetrics
      tags: [dashboard]
      summary: CRM dashboard metrics
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /v1/admin/dashboard:
    get:
      operationId: adminDashboard
      tags: [admin]
      summary: Platform admin dashboard metrics
      responses:
        "200":
          description: Admin dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboardData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/organizations:
    get:
      operationId: adminListOrganizations
      tags: [admin]
      summary: List all organizations
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrganizationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Search platform users
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/roles:
    get:
      operationId: adminListRoles
      tags: [admin]
      summary: List platform roles
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRoleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/permissions:
    get:
      operationId: adminListPermissions
      tags: [admin]
      summary: List platform permissions
      responses:
        "200":
          description: Permission list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPermissionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/audit-logs:
    get:
      operationId: adminSearchAuditLogs
      tags: [admin]
      summary: Search platform audit logs
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
        - name: entity
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: string
        - name: actorUserId
          in: query
          schema:
            type: string
        - name: requestId
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditLogListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/feature-flags:
    get:
      operationId: adminListFeatureFlags
      tags: [admin]
      summary: List feature flags
      responses:
        "200":
          description: Feature flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFeatureFlagListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/admin/feature-flags/{key}:
    patch:
      operationId: adminUpdateFeatureFlag
      tags: [admin]
      summary: Update a global feature flag
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminFeatureFlagUpdateRequest"
      responses:
        "200":
          description: Updated feature flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminFeatureFlagRow"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
...
    Unauthorized:
      description: Authentication is required, or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: >
        Authenticated, but lacking the required permission or tenant scope.
        Retrying with the same credentials will not succeed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: >
        The resource does not exist, or exists outside the caller's tenant.
        The two are deliberately indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimited:
      description: Too many requests. Retry after the window resets.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >
        An unexpected fault. The message is generic by design; quote
        `error.requestId` when reporting it.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: The service is temporarily unable to handle the request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DashboardMetrics:
      type: object
      properties:
        newLeads:
          type: number
        qualified:
          type: number
        won:
          type: number
        lost:
          type: number
        openLeads:
          type: number
        averageResponseMinutes:
          type: number
          nullable: true
      required:
        - newLeads
        - qualified
        - won
        - lost
        - openLeads
...
    FieldError:
      type: object
      description: A single field-level validation failure.
      properties:
        path:
          type: string
          description: Dotted path to the offending field, e.g. `contact.email`.
        message:
          type: string
      required:
        - path
        - message
...
    AdminDashboardData:
      type: object
      properties:
        organizationCount:
          type: integer
        userCount:
          type: integer
        activeUserCount:
          type: integer
        auditLogCount24h:
          type: integer
        openOutboxJobs:
          type: integer
      required:
        - organizationCount
        - userCount
        - activeUserCount
        - auditLogCount24h
        - openOutboxJobs
    AdminOrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        status:
          type: string
        type:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - status
        - type
        - createdAt
    AdminOrganizationListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrganizationSummary"
      required:
        - items
    AdminUserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
        roles:
          type: array
          items:
            type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - status
        - roles
    AdminUserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
      required:
        - items
    AdminRoleSummary:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isSystem:
          type: boolean
      required:
        - id
        - key
        - name
        - isSystem
    AdminRoleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminRoleSummary"
      required:
        - items
    AdminPermissionSummary:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        domain:
          type: string
      required:
        - id
        - key
        - name
        - domain
    AdminPermissionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminPermissionSummary"
      required:
        - items
    PlatformAuditLogEntry:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: string
          nullable: true
        entity:
          type: string
        entityId:
          type: string
        action:
          type: string
        eventType:
          type: string
        severity:
          type: string
        actorUserId:
          type: string
          nullable: true
        actorKind:
          type: string
        requestId:
          type: string
        occurredAt:
          type: string
          format: date-time
        previousValues:
          nullable: true
        newValues:
          nullable: true
      required:
        - id
        - entity
        - entityId
        - action
        - eventType
        - severity
        - actorKind
        - requestId
        - occurredAt
    AdminAuditLogListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlatformAuditLogEntry"
      required:
        - items
    AdminFeatureFlagRow:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
      required:
        - id
        - key
        - enabled
    AdminFeatureFlagListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminFeatureFlagRow"
      required:
        - items
    AdminFeatureFlagUpdateRequest:
      type: object
      properties:
        enabled:
          type: boolean
        description:
          type: string
          nullable: true
      required:
        - enabled
...
    ErrorResponse:
      type: object
      description: >
        Standard error envelope returned by every failing endpoint. The `code`
        enum is kept in lockstep with `ErrorCode` in
        `lib/platform/src/errors`; adding a code requires updating both.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: >
                Stable, machine-readable classification. Clients branch on this
                rather than on the HTTP status or the message text.
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - unprocessable_entity
                - rate_limited
                - internal_error
                - service_unavailable
            message:
              type: string
              description: >
                Human-readable summary. Safe to display for expected errors;
                generic for internal faults, which never disclose their cause.
            details:
              type: object
              description: >
                Structured, client-safe detail. Carries `fieldErrors` for
                validation failures; otherwise error-specific.
              properties:
                fieldErrors:
                  type: array
                  items:
                    $ref: "#/components/schemas/FieldError"
              additionalProperties: true
            requestId:
              type: string
              description: >
                Correlation id, matching the `x-request-id` response header and
                the server logs. Quote it when reporting a fault.
          required:
            - code
            - message
      required:
        - error